pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.2.0"
description = ""
authors = ["Zhdanovich Timofey <zhdanovich.tv@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torch = "^2.1.0"
scikit-learn = "^1.3.1"
torchvision = "^0.16.0"
numpy = "^1.26.0"
pandas = "^2.1.1"

[tool.poetry.group.dev.dependencies]
dvc-gdrive = "^2.20.0"
dvc = "^3.26.0"
pre-commit = "^3.4.0"
fire = "^0.5.0"
hydra-core = "^1.3"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["somepytools"]
profile = "black"
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
